# Conflicts:
#	installer/manifests/keptn/charts/control-plane/values.yaml

+++ configuration-service/restapi/operations/service_resource/get_project_project_name_stage_stage_name_service_service_name_resource_parameters.go
@@ -117,7 +117,6 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) Bi
 	if err := o.bindStageName(rStageName, rhkStageName, route.Formats); err != nil {
 		res = append(res, err)
 	}
-
 	if len(res) > 0 {
 		return errors.CompositeValidationError(res...)
 	}
@@ -133,6 +132,7 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: false
 	// AllowEmptyValue: false
+
 	if raw == "" { // empty values pass all other validations
 		// Default values have been previously initialized by NewGetProjectProjectNameStageStageNameServiceServiceNameResourceParams()
 		return nil
@@ -156,10 +156,10 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: false
 	// AllowEmptyValue: false
+
 	if raw == "" { // empty values pass all other validations
 		return nil
 	}
-
 	o.NextPageKey = &raw
 
 	return nil
@@ -174,6 +174,7 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: false
 	// AllowEmptyValue: false
+
 	if raw == "" { // empty values pass all other validations
 		// Default values have been previously initialized by NewGetProjectProjectNameStageStageNameServiceServiceNameResourceParams()
 		return nil
@@ -195,11 +196,11 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 // validatePageSize carries on validations for parameter PageSize
 func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) validatePageSize(formats strfmt.Registry) error {
 
-	if err := validate.MinimumInt("pageSize", "query", int64(*o.PageSize), 1, false); err != nil {
+	if err := validate.MinimumInt("pageSize", "query", *o.PageSize, 1, false); err != nil {
 		return err
 	}
 
-	if err := validate.MaximumInt("pageSize", "query", int64(*o.PageSize), 50, false); err != nil {
+	if err := validate.MaximumInt("pageSize", "query", *o.PageSize, 50, false); err != nil {
 		return err
 	}
 
@@ -215,7 +216,6 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: true
 	// Parameter is provided by construction from the route
-
 	o.ProjectName = raw
 
 	return nil
@@ -230,7 +230,6 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: true
 	// Parameter is provided by construction from the route
-
 	o.ServiceName = raw
 
 	return nil
@@ -245,7 +244,6 @@ func (o *GetProjectProjectNameStageStageNameServiceServiceNameResourceParams) bi
 
 	// Required: true
 	// Parameter is provided by construction from the route
-
 	o.StageName = raw
 
 	return nil
